Modular sub-project based repository structure to be able to add Freegeoip and MaxMind dependencies one at a time.
