+++ .travis.yml
@@ -1,7 +1,6 @@
 language: scala
 script:
    - sbt clean compile test publishLocal
-   - cd sample/java && sbt clean compile test
 
 jdk:
    - oraclejdk8
