--- conf/reference.conf
@@ -1,10 +1,0 @@
-
-play.modules.enabled += com.edulify.modules.geolocation.GeolocationModule
-
-geolocation {
-  cache {
-    on = true
-    ttl = 100
-  }
-  maxmind.license = "abc123"
-}
