+++ geolocation/src/main/resources/reference.conf
@@ -1,0 +1,4 @@
+geolocation.cache {
+  on = true
+  ttl = 100
+}
