+++ maxmind-geoip2web-guice/src/main/resources/reference.conf
@@ -1,0 +1,1 @@
+geolocation.maxmind.license = ~
