+++ sample/java/conf/application.conf
@@ -11,4 +11,7 @@ play.http.secret.key="i[7Ve:r2mSr`PKm1x@kVYm0XfqW[vLarVbjGuP/Q8I8p@iZ8RLVPSiq@6Z
 # ~~~~~
 play.i18n.langs=["en"]
 
-play.modules.enabled += "com.edulify.modules.geolocation.providers.FreegeoipModule"
+play.modules {
+  enabled += "com.edulify.modules.geolocation.FreegeoipModule"
+  enabled += "com.edulify.modules.ws.standalone.StandaloneAhcWSModule"
+}
