+++ version.sbt
@@ -1,0 +1,1 @@
+version in ThisBuild := "2.1.0"
